[HUGE CHANGE] Changed the permissions system

+++ taiga/projects/issues/api.py
@@ -15,18 +15,20 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from django.utils.translation import ugettext_lazy as _
+from django.shortcuts import get_object_or_404
 from django.db.models import Q
+from django.http import Http404
 
-from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework import filters
 
 from taiga.base import filters
 from taiga.base import exceptions as exc
-from taiga.base.decorators import list_route, detail_route
-from taiga.base.api import ModelCrudViewSet
+from taiga.base.decorators import detail_route
+from taiga.base.api import ModelCrudViewSet, ModelListViewSet
 from taiga.base import tags
+
+from taiga.users.models import User
 
 from taiga.projects.notifications import WatchedResourceMixin
 from taiga.projects.occ import OCCResourceMixin
@@ -42,7 +44,7 @@ from . import serializers
 
 
 class IssuesFilter(filters.FilterBackend):
-    filter_fields = ( "status", "severity", "priority", "owner", "assigned_to", "tags", "type")
+    filter_fields = ("status", "severity", "priority", "owner", "assigned_to", "tags", "type")
     _special_values_dict = {
         'true': True,
         'false': False,
@@ -80,7 +82,7 @@ class IssuesFilter(filters.FilterBackend):
 
         for name, value in filter(lambda x: x[0] != "tags", filterdata.items()):
             if None in value:
-                qs_in_kwargs = {"{0}__in".format(name): [v for v in value if v != None]}
+                qs_in_kwargs = {"{0}__in".format(name): [v for v in value if v is not None]}
                 qs_isnull_kwargs = {"{0}__isnull".format(name): True}
                 queryset = queryset.filter(Q(**qs_in_kwargs) | Q(**qs_isnull_kwargs))
             else:
@@ -104,9 +106,9 @@ class IssuesOrdering(filters.FilterBackend):
 class IssueViewSet(OCCResourceMixin, HistoryResourceMixin, WatchedResourceMixin, ModelCrudViewSet):
     serializer_class = serializers.IssueNeighborsSerializer
     list_serializer_class = serializers.IssueSerializer
-    permission_classes = (IsAuthenticated, permissions.IssuePermission)
+    permission_classes = (permissions.IssuePermission, )
 
-    filter_backends = (filters.IsProjectMemberFilterBackend, IssuesFilter, IssuesOrdering)
+    filter_backends = (filters.CanViewIssuesFilterBackend, IssuesFilter, IssuesOrdering)
     retrieve_exclude_filters = (IssuesFilter,)
 
     filter_fields = ("project",)
@@ -133,42 +135,65 @@ class IssueViewSet(OCCResourceMixin, HistoryResourceMixin, WatchedResourceMixin,
     def pre_conditions_on_save(self, obj):
         super().pre_conditions_on_save(obj)
 
-        if (obj.project.owner != self.request.user and
-                obj.project.memberships.filter(user=self.request.user).count() == 0):
-            raise exc.PermissionDenied(_("You don't have permissions for add/modify this issue."))
-
         if obj.milestone and obj.milestone.project != obj.project:
-            raise exc.PermissionDenied(_("You don't have permissions for add/modify this issue."))
+            raise exc.PermissionDenied(_("You don't have permissions to set this milestone to this issue."))
 
         if obj.status and obj.status.project != obj.project:
-            raise exc.PermissionDenied(_("You don't have permissions for add/modify this issue."))
+            raise exc.PermissionDenied(_("You don't have permissions to set this status to this issue."))
 
         if obj.severity and obj.severity.project != obj.project:
-            raise exc.PermissionDenied(_("You don't have permissions for add/modify this issue."))
+            raise exc.PermissionDenied(_("You don't have permissions to set this severity to this issue."))
 
         if obj.priority and obj.priority.project != obj.project:
-            raise exc.PermissionDenied(_("You don't have permissions for add/modify this issue."))
+            raise exc.PermissionDenied(_("You don't have permissions to set this priority to this issue."))
 
         if obj.type and obj.type.project != obj.project:
-            raise exc.PermissionDenied(_("You don't have permissions for add/modify this issue."))
+            raise exc.PermissionDenied(_("You don't have permissions to set this type to this issue."))
 
-    @detail_route(methods=['post'], permission_classes=(IsAuthenticated,))
+    @detail_route(methods=['post'])
     def upvote(self, request, pk=None):
-        issue = self.get_object()
+        issue = get_object_or_404(models.Issue, pk=pk)
+
+        self.check_permissions(request, 'upvote', issue)
+
         votes_service.add_vote(issue, user=request.user)
         return Response(status=status.HTTP_200_OK)
 
-    @detail_route(methods=['post'], permission_classes=(IsAuthenticated,))
+    @detail_route(methods=['post'])
     def downvote(self, request, pk=None):
-        issue = self.get_object()
+        issue = get_object_or_404(models.Issue, pk=pk)
+
+        self.check_permissions(request, 'downvote', issue)
+
         votes_service.remove_vote(issue, user=request.user)
         return Response(status=status.HTTP_200_OK)
 
 
-class VotersViewSet(ModelCrudViewSet):
+class VotersViewSet(ModelListViewSet):
     serializer_class = votes_serializers.VoterSerializer
     list_serializer_class = votes_serializers.VoterSerializer
-    permission_classes = (IsAuthenticated,)
+    permission_classes = (permissions.IssueVotersPermission, )
+
+    def retrieve(self, request, *args, **kwargs):
+        pk = kwargs.get("pk", None)
+        issue_id = kwargs.get("issue_id", None)
+        issue = get_object_or_404(models.Issue, pk=issue_id)
+
+        self.check_permissions(request, 'retrieve', issue)
+
+        try:
+            self.object = votes_service.get_voters(issue).get(pk=pk)
+        except User.DoesNotExist:
+            raise Http404
+
+        serializer = self.get_serializer(self.object)
+        return Response(serializer.data)
+
+    def list(self, request, *args, **kwargs):
+        issue_id = kwargs.get("issue_id", None)
+        issue = get_object_or_404(models.Issue, pk=issue_id)
+        self.check_permissions(request, 'list', issue)
+        return super().list(request, *args, **kwargs)
 
     def get_queryset(self):
         issue = models.Issue.objects.get(pk=self.kwargs.get("issue_id"))
